Upload the embed's photo through the registration endpoint
One endpoint rather than two: an embed registration's b_param belongs to
the organization handing out the form, so find_token won't hand it to
whoever fills the form in - the organization's own rule, that holding
the token is enough, is the fallback.

A second endpoint would also have been a second budget for the same
anonymous b_params, since with_organization_or_no_creator matches a
creator-less registration too. The one hourly budget goes to 60: a
shop's embed form registers bike after bike from the one address.
