+++ app/views/organizations/_embed_fields.html.erb
@@ -381,6 +381,6 @@
     %>
     <%= render UI::Forms::FileUpload::Component.new(form_builder: f, attribute: :image,
           accept: ImageUploader.permitted_extensions,
-          direct_upload_url: (embed_direct_uploads_path(b_param_token: @b_param.id_token) if @b_param.id_token.present?)) %>
+          direct_upload_url: (register_direct_uploads_path(b_param_token: @b_param.id_token) if @b_param.id_token.present?)) %>
   <% end %>
 </div>
