+++ config/initializers/rack_attack.rb
@@ -11,10 +11,8 @@ class Rack::Attack
   CSP_REPORTS_PATH = "/csp_reports"
   DIRECT_UPLOADS_PATH = "/rails/active_storage/direct_uploads"
   REGISTER_DIRECT_UPLOADS_PATH = "/register/direct_uploads"
-  EMBED_DIRECT_UPLOADS_PATH = "/embed/direct_uploads"
   DIRECT_UPLOAD_MAX = ENV.fetch("RACK_ATTACK_DIRECT_UPLOAD_LIMIT", 20).to_i
-  REGISTER_DIRECT_UPLOAD_MAX = ENV.fetch("RACK_ATTACK_REGISTER_DIRECT_UPLOAD_LIMIT", 10).to_i
-  EMBED_DIRECT_UPLOAD_MAX = ENV.fetch("RACK_ATTACK_EMBED_DIRECT_UPLOAD_LIMIT", 60).to_i
+  REGISTER_DIRECT_UPLOAD_MAX = ENV.fetch("RACK_ATTACK_REGISTER_DIRECT_UPLOAD_LIMIT", 60).to_i
 
   SENSITIVE_AUTH_PATHS = %w[
     /session/create_magic_link
@@ -84,21 +82,17 @@ class Rack::Attack
     request.ip if request.patch? && request.path == "/my_account"
   end
 
-  # Each direct upload hands out a presigned URL to write into our bucket. Every endpoint
-  # verifies who's asking in the app - signed in for one, a registration token for the others -
-  # so these only cap how fast one address can ask. A registration needs a single upload, hence
-  # the far tighter hourly budget on the anonymous one; an embed form is anonymous too, but a
-  # shop registers bike after bike from the one address.
+  # Each direct upload hands out a presigned URL to write into our bucket. Both endpoints
+  # verify who's asking in the app - signed in for one, a registration token for the other -
+  # so these only cap how fast one address can ask. One registration needs a single upload, but
+  # a shop's embed form registers bike after bike from the one address, hence the hourly rather
+  # than per-minute budget on the anonymous one.
   throttle("direct_uploads/ip", limit: DIRECT_UPLOAD_MAX, period: 1.minute) do |request|
     request.ip if request.post? && request.path == DIRECT_UPLOADS_PATH
   end
 
   throttle("register_direct_uploads/ip", limit: REGISTER_DIRECT_UPLOAD_MAX, period: 1.hour) do |request|
     request.ip if request.post? && request.path == REGISTER_DIRECT_UPLOADS_PATH
-  end
-
-  throttle("embed_direct_uploads/ip", limit: EMBED_DIRECT_UPLOAD_MAX, period: 1.hour) do |request|
-    request.ip if request.post? && request.path == EMBED_DIRECT_UPLOADS_PATH
   end
 
   self.throttled_responder = lambda do |request|
